Move query classes used in streaming search to separate sub-folder and sub-library.
