+++ searchlib/src/tests/attribute/imported_search_context/imported_search_context_test.cpp
@@ -4,7 +4,7 @@
 #include <vespa/searchlib/attribute/attribute_read_guard.h>
 #include <vespa/searchlib/attribute/imported_search_context.h>
 #include <vespa/searchlib/fef/termfieldmatchdata.h>
-#include <vespa/searchlib/query/queryterm.h>
+#include <vespa/searchlib/query/query_term_ucs4.h>
 #include <vespa/searchlib/queryeval/simpleresult.h>
 #include <vespa/searchlib/test/imported_attribute_fixture.h>
 #include <vespa/vespalib/test/insertion_operators.h>
