+++ searchlib/src/vespa/searchlib/query/query_term_decoder.cpp
@@ -1,7 +1,7 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "query_term_decoder.h"
-#include "query.h"
+#include <vespa/searchlib/query/streaming/query.h>
 #include <vespa/vespalib/util/exceptions.h>
 
 namespace search {
