+++ searchlib/src/vespa/searchlib/query/streaming/CMakeLists.txt
@@ -1,0 +1,9 @@
+# Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+vespa_add_library(searchlib_query_streaming OBJECT
+    SOURCES
+    query.cpp
+    querynode.cpp
+    querynoderesultbase.cpp
+    queryterm.cpp
+    DEPENDS
+)
