+++ searchlib/src/vespa/searchlib/query/streaming/querynoderesultbase.h
