+++ searchlib/src/vespa/searchlib/query/streaming/queryterm.cpp
